Minor rework and cleanup

+++ Void/src/Void/Platform/OpenGL/Components/VertexBuffer/OpenGLVertexBuffer.cpp
@@ -6,7 +6,7 @@
 namespace Void::Rendering {
     
     OpenGLVertexBuffer::OpenGLVertexBuffer(const void* vertices, const size_t size)
-        : m_Count(size), m_Data(nullptr)
+        : m_Count((uint32_t)size), m_Data(nullptr)
     {
         GLCall(glGenBuffers(1, &m_ID));
         GLCall(glBindBuffer(GL_ARRAY_BUFFER, m_ID));
